Avoiding duplicate test cases in template and param test cases
- fixing logging on non-existant stream
- fixing doc test cases with duplicate test cases
- updating log and report tests

+++ test/test-organization-ts/test-tree-management-test.cpp
@@ -91,7 +91,7 @@ struct F2 {
 
 BOOST_FIXTURE_TEST_SUITE( S5, F2 )
 
-typedef boost::mpl::vector<int,float,char,double,int> test_types;
+typedef boost::mpl::vector<int,float,char,double,short int> test_types;
 BOOST_AUTO_TEST_CASE_TEMPLATE( tctempl, T, test_types )
 {
 }
